requirements.txt
python-dotenv
python-telegram-bot
asyncio
pymongo
motor
